\name{subgroupTEtest-package}
\alias{subgroupTEtest-package}

\docType{package}
\title{
\packageTitle{subgroupTEtest}
}
\description{
\packageDescription{subgroupTEtest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{subgroupTEtest}
\packageIndices{subgroupTEtest}
The package implements the tests proposed in "An Omnibus Test for Detection of SubgroupTreatment Effects via Data Partitioning".
The test for censored data depends on the package "faft", which implements the fast censored linear regression in Huang, Y. (2013) Fast censored linear regression. Scandinavian Journal of Statistics 40, 789--806. The package "faft" is not available on CRAN but can be downloaded from Dr. Huang's personal website.
}
\author{
\packageAuthor{subgroupTEtest}

Maintainer: \packageMaintainer{subgroupTEtest}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, ~~
~~ one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
