% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subgroupTEtest}
\alias{subgroupTEtest}
\title{Test for uncensored data}
\usage{
subgroupTEtest(xu, y, tr, fx, g, B)
}
\arguments{
\item{xu}{A matrix of partitions: each column corresponds to a binary partition and each row 
corresponds to an observation. 
The elements in the matrix must be 1 (group 1) or 2 (group 2). 
The element in the ith row and jth column indicates the group membership of the ith observation
based on the jth partition.}

\item{y}{A numeric vector containing the outcome.}

\item{tr}{A binary vector containing the treatment indicator: 1 - treatment, 0 - control.}

\item{fx}{A numeric matrix of covariates to be used in f.}

\item{g}{A vector containing the values of g.}
}
\value{
The p value of the test.
}
\description{
Test for uncensored data
}
\examples{
library(glmnet)
set.seed(1)
N <- 800
P <- 50
beta <- rep(0,length = P)
beta[1] <- 1
beta[P] <- -1
X <- matrix(rnorm(N*P), ncol = P)
TR <- rbinom(N, 1, 0.5)
Y <-  X \%*\% beta + TR*0.25*(2*(X[,1]>0 | X[,25]>0)-1)+ rnorm(N, sd = 0.5)
  
# partitions
X2 <- apply(X,2,function(u) (u<0)+1)
# select variables in f
flasso <- glmnet(y = Y, x = X)
fbeta <- coef(flasso, s = 0.025)
X3 <- X[,fbeta[2:(P+1)]!=0]
# calculate g
glasso <- glmnet(y = Y, x = cbind(X,TR*X,TR))
gbeta <- coef(glasso, s = 0.05)[(P+2):(2*P+2)]
g <- cbind(TR*X,TR) \%*\% gbeta
# p-value
subgroupTEtest(X2,Y,TR,X3,g)
}
