% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subgroupTEtestSurv}
\alias{subgroupTEtestSurv}
\title{Test for censored data. The functions f and g use the same pre-specified set of covariates 
and are estimated from working AFT models. Note that the follow-up time is on the log scale.}
\usage{
subgroupTEtestSurv(x, xu, y, dlt, tr, B)
}
\arguments{
\item{x}{A numeric matrix of covariates to be used in f and g.}

\item{xu}{A matrix of partitions: each column corresponds to a binary partition and each row 
corresponds to an observation. 
The elements in the matrix must be 1 (group 1) or 2 (group 2). 
The element in the ith row and jth column indicates the group membership of the ith observation
based on the jth partition.}

\item{y}{A numeric vector containing the follow-up time on the log scale.}

\item{dlt}{A binary vector containing the event indicator: 1 - event, 0 - censored.}

\item{tr}{A binary vector containing the treatment indicator: 1 - treatment, 0 - control.}

\item{B}{An integer of the number of bootstrap samples}
}
\value{
The p value of the test.
}
\description{
Test for censored data. The functions f and g use the same pre-specified set of covariates 
and are estimated from working AFT models. Note that the follow-up time is on the log scale.
}
\examples{
set.seed(2)
N <- 800
beta <- rep(0,length = 50)
beta[1] <- 1
beta[50] <- -1
X <- matrix(rnorm(N*50), ncol = 50)
TR <- rbinom(N, 1, 0.5)
Y0 <-  X \%*\% beta + TR*0.5*(2*(X[,1]>0 | X[,25]>0)-1)+ rnorm(N, sd = 0.5)
Y0 <- as.vector(Y0)
C <- log(rexp(N, 0.1))
Y <- pmin(Y0,C)
DLT <- Y0 <= C
X2 <- apply(X,2,function(u) (u<0)+1)
# p-value
subgroupTEtestSurv(X[,c(1,25,50)], X2, Y, DLT, TR, 1000)
subgroupTEtestSurv(X[,c(1,50)], X2, Y, DLT, TR, 1000)
}
