% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subgroupTEtestSurv_g}
\alias{subgroupTEtestSurv_g}
\title{Test for censored data. Users can specify the values of the g function for each observation. 
The f function is estimated using a working AFT model. Note that the follow-up time is on the log scale.}
\usage{
subgroupTEtestSurv_g(x, xu, y, dlt, tr, g, B)
}
\arguments{
\item{x}{A numeric matrix of covariates to be used in f.}

\item{xu}{A matrix of partitions: each column corresponds to a binary partition and each row 
corresponds to an observation. 
The elements in the matrix must be 1 (group 1) or 2 (group 2). 
The element in the ith row and jth column indicates the group membership of the ith observation
based on the jth partition.}

\item{y}{A numeric vector containing the follow-up time on the log scale.}

\item{dlt}{A binary vector containing the event indicator: 1 - event, 0 - censored.}

\item{tr}{A binary vector containing the treatment indicator: 1 - treatment, 0 - control.}

\item{g}{A vector containing the values of g.}

\item{B}{An integer of the number of bootstrap samples}
}
\value{
The p value of the test.
}
\description{
Test for censored data. Users can specify the values of the g function for each observation. 
The f function is estimated using a working AFT model. Note that the follow-up time is on the log scale.
}
\examples{
set.seed(2)
N <- 800
beta <- rep(0,length = 50)
beta[1] <- 1
beta[P] <- -1
X <- matrix(rnorm(N*P), ncol = P)
TR <- rbinom(N, 1, 0.5)
Y0 <-  X \%*\% beta + TR*0.25*(2*(X[,1]>0 | X[,25]>0)-1)+ rnorm(N, sd = 0.5)
Y0 <- as.vector(Y0)
C <- log(rexp(N, 0.1))
Y <- pmin(Y0,C)
DLT <- Y0 <= C
X2 <- apply(X,2,function(u) (u<0)+1)
X1 <- X[,c(1,25,50)]
# p-value (g = 0)
subgroupTEtestSurv_g(X1, X2, Y, DLT, TR, rep(0,N), 1000)

# Estimate g from a working AFT model
library(faft)
fitg <- faft(Y, DLT, cbind(X1,TR*X1,TR))
G <- cbind(TR*X1,TR) \%*\% fitg$beta[4:7]
# p value
subgroupTEtestSurv_g(X1, X2, Y, DLT, TR, G, 1000)
}
