% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subgroupTEtest}
\alias{subgroupTEtest}
\title{Test for uncensored data}
\usage{
subgroupTEtest(xu, y, tr, fx, g, B)
}
\arguments{
\item{xu}{A matrix of partitions: each column corresponds to a binary partition and each row 
corresponds to an observation. 
The elements in the matrix must be 1 (group 1) or 2 (group 2). 
The element in the ith row and jth column indicates the group membership of the ith observation
based on the jth partition.}

\item{y}{A numeric vector containing the outcome.}

\item{tr}{A binary vector containing the treatment indicator: 1 - treatment, 0 - control.}

\item{fx}{A numeric matrix of covariates to be used in f.}

\item{g}{A vector containing the values of g.}

\item{B}{An integer of the number of bootstrap samples.}
}
\value{
The p value of the test.
}
\description{
Test for uncensored data
}
\examples{
set.seed(1)
N <- 800
beta <- rep(0,length = 50)
beta[1] <- 1
beta[50] <- -1
X <- matrix(rnorm(N*50), ncol = 50)
TR <- rbinom(N, 1, 0.5)
Y <-  X \%*\% beta + TR*0.5*(2*(X[,1]>0 | X[,25]>0)-1)+ rnorm(N, sd = 0.5)
  
# partition matrix
X2 <- apply(X,2,function(u) (u<0)+1)
# covariates in f
X1 <- X[,c(1,25,50)]
# p-value (g = 0)
subgroupTEtest(X2,Y,TR,X1,rep(0,N),1000)

# estimate g from a working linear model
fitg <- lm(Y~X1*TR)
G <- cbind(TR,TR*X1) \%*\% coef(fitg)[5:8]
subgroupTEtest(X2,Y,TR,X1,G,1000)
}
